+++ ClassicLootManager/Modules/Auctioning/AuctionManager/AuctionManager.lua
@@ -1093,9 +1093,14 @@ local function ValidateBid(auction, item, name, userResponse)
             if auction:GetAlwaysAllowAllInBids() and (current == value) then
                 return true
             end
-            -- always allow bidding min in ascending mode but ONLY if player haven't bid yet
-            if value == values[CONSTANTS.SLOT_VALUE_TIER.BASE] and auction:GetAlwaysAllowBaseBids() and not auction:HasBids(item:GetItemID(), name) then
-                return true
+            -- always allow bidding base or 0 in ascending mode but ONLY if player haven't bid yet
+            if not auction:HasBids(item:GetItemID(), name) then
+                if value == values[CONSTANTS.SLOT_VALUE_TIER.BASE] and auction:GetAlwaysAllowBaseBids() then
+                    return true
+                end
+                if (value == 0) and auction:GetAlwaysAllowAll0() then
+                    return true
+                end
             end
             if value == item:GetHighestBid() then
                 if auction:GetAllowEqualBids() then